[FIX] Button New Blank not move when undo bar presents [FIX] Delete player button not match with dark theme [IMPLEMENT] Start game with initial score in preference [IMPLEMENT] Initialize Help & feedback activity

+++ app/src/main/java/com/tnt/scoreboard/dataAccess/PlayerDAO.java
@@ -32,7 +32,7 @@ public class PlayerDAO extends BaseDAO<Player> {
                 cursor.getLong(cursor.getColumnIndexOrThrow(Player.COLUMN_ID)),
                 cursor.getLong(cursor.getColumnIndexOrThrow(Player.COLUMN_GAME_ID)),
                 cursor.getString(cursor.getColumnIndexOrThrow(Player.COLUMN_NAME)),
-                cursor.getDouble(cursor.getColumnIndexOrThrow(Player.COLUMN_SCORE)),
+                cursor.getLong(cursor.getColumnIndexOrThrow(Player.COLUMN_SCORE)),
                 cursor.getLong(cursor.getColumnIndexOrThrow(Player.COLUMN_COLOR))
         );
     }
